+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## - 0.0.4
 - Implement: JOIN and WHERE to QueryBuilder and API
 - Typo fix: in API() tokken => token 
-- Typo fix: in API() erros => errors 
+- Typo fix: in API() erros => errors
+- Merge: request + setup.php
 
 ## < 0.0.4
 - Tu much, show in the commits